Update inventeringtrial.py and bugfixes
Made loop for all the 80 placeholders. 

Fixed so deleting entry doesn't redo the search function unmodified, which made it look as if the database got deleted. Now it requerys but removes the unique identifier from it so it disappears from the search results. 

Same thing but for editing an entry. Editing now updates the search results to reflect the database change. 

Added some comments.